Troops Evacuation
The British and The French troops are stranded on the sea shore at Dunkirk. There are n piers along the sea bank in a row, on some of which a number of soldiers wait for evacuation. The i'th pier has either i or no soldiers waiting on it, i e The 6th pier may have either 6 or 0 soldiers on it . Mr Dawson, a brave sailor from England is on a mission to save the soldiers with his ship. But he can anchor his ship at only one pier at a time and save all the soldiers on it, and then move on to any of the other piers. 
Much to his dismay, as soon as he anchors on a pier, the German bomber planes sense his proximity. Thankfully, they are guided by a malfunctioning radar system. So whenever Mr Dawson anchors on a pier m, the bombers arrive and drop bombs on both the piers adjacent to m, i.e. pier m-1 and pier m+1, killing any and all soldiers on either of the two piers, but the soldiers at pier m are saved successfully. Given the list of piers on which their are any soldiers, what is the maximum number of soldiers Mr Dawson can save ? 
Input: 
One line of integers representing the index(1 <= index <= n) of piers which have soldiers separated by space. 
Output: 
A single integer representing The maximum number of soldiers that can be saved. 
Sample Input 1: 
8 12 7 3 
Sample Output 1: 
23 
Explanation: If Dawsons picks pier 12, the planes bomb 11 and 13, which do not have any soldiers. So no lives lost Then if he pic. 8, the planes bomb 7 and 9, of which 7 exists in The list. So only 7 soldiers die.

Sample Input 2: 
879103 
Sample Output 2: 
21 
Explanation: If Dawson picks 10, he loses 9 soldiers and then loses 7 more when he picks 8. 
Constraints 
0 < n < 100000