Girls in tech Hackathon
The Girls in Tech Hackathon is a code-a-thon where developers, designers, scientists, students, entrepreneurs and educators gather to collaborate on prmects including applications, software, hardware, data visualization and platform solutions.

 The Participants are sitting around a round table, a. they are numbered from team 1 to team n in the clockwise direction.This means that the Participant are numbered 1,2...,n-1,n and participants 1 and n are sitting next to each other. After the Hackathon duration, judges started reviewing the performance of each participant. However, some of participants haven't finished the project. Each participant needs t, extra minutes to complete their Project. Judges started reviewing the projects sequentially in the clockwise direction, starting with team x, and it takes them exactly 1 minute to review each project. This means the team x gets no extra time to work, whereas team x+1 gets 1 min more to work and so on. Even if the project is not completed by participant, but still the judges spend 1 min to understand the idea and rank the participant. 
 
Input Format: 
The first line contains a single positive integer, n, denoting the number of participants in the hackathon. 
given the values of t1, t2, t3...tn extra time requires by each participant to complete the project. You have to guide judges in selecting the best team x to start reviewing the project so that number of participants able to complete their projects are maximal. 

Output Format: 
Print x on a new line. If ere are multiple such teams, select the smallest one.